requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-bootstrap4==23.2
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
packaging==23.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.4
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-bootstrap4==23.2
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
packaging==23.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.4
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
coverage==7.3.0
Django==4.2.4
django-bootstrap4==23.2
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
packaging==23.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.4
